+++ flows/group.go
@@ -43,6 +43,10 @@ func (g *Group) CheckDynamicMembership(env envs.Environment, contact *Contact) (
 		panic("can't check membership on a non-dynamic group")
 	}
 
+	if contact.Blocked() || contact.Stopped() {
+		return false, nil
+	}
+
 	return contactql.EvaluateQuery(env, g.parsedQuery, contact)
 }