+++ flows/services_test.go
@@ -16,19 +16,25 @@ func TestHTTPLogs(t *testing.T) {
 
 	httpx.SetRequestor(httpx.NewMockRequestor(map[string][]httpx.MockResponse{
 		"http://temba.io/": []httpx.MockResponse{
-			httpx.NewMockResponse(200, nil, "hello", 1),
-			httpx.NewMockResponse(400, nil, "is error", 1),
+			httpx.NewMockResponse(200, nil, "hello"),
+			httpx.NewMockResponse(400, nil, "is error"),
 			httpx.MockConnectionError,
 		},
 	}))
 
-	trace1, err := httpx.NewTrace(http.DefaultClient, "GET", "http://temba.io/", nil, nil, nil, nil)
+	req1, err := httpx.NewRequest("GET", "http://temba.io/", nil, nil)
+	require.NoError(t, err)
+	trace1, err := httpx.DoTrace(http.DefaultClient, req1, nil, nil, -1)
 	require.NoError(t, err)
 
-	trace2, err := httpx.NewTrace(http.DefaultClient, "GET", "http://temba.io/", nil, nil, nil, nil)
+	req2, err := httpx.NewRequest("GET", "http://temba.io/", nil, nil)
+	require.NoError(t, err)
+	trace2, err := httpx.DoTrace(http.DefaultClient, req2, nil, nil, -1)
 	require.NoError(t, err)
 
-	trace3, err := httpx.NewTrace(http.DefaultClient, "GET", "http://temba.io/", nil, nil, nil, nil)
+	req3, err := httpx.NewRequest("GET", "http://temba.io/", nil, nil)
+	require.NoError(t, err)
+	trace3, err := httpx.DoTrace(http.DefaultClient, req3, nil, nil, -1)
 	require.EqualError(t, err, "unable to connect to server")
 
 	log1 := flows.NewHTTPLog(trace1, flows.HTTPStatusFromCode)