+++ services/webhooks/service.go
@@ -43,21 +43,23 @@ func (s *service) Call(session flows.Session, request *http.Request) (*flows.Web
 		}
 	}
 
+	// If user has explicitly set Accept-Encoding: gzip, remove it as Transport will add this itself,
+	// and it only does automatic decompression if its the one to set it.
+	if request.Header.Get("Accept-Encoding") == "gzip" {
+		request.Header.Del("Accept-Encoding")
+	}
+
 	trace, err := httpx.DoTrace(s.httpClient, request, s.httpRetries, s.httpAccess, s.maxBodyBytes)
 	if trace != nil {
 		call := &flows.WebhookCall{Trace: trace}
-
-		// for webhook calls, we're only interested in valid JSON response bodies
-		if len(trace.ResponseBody) > 0 && !json.Valid(trace.ResponseBody) {
-			call.ResponseBody = nil
-			call.BodyIgnored = true
-		}
 
 		// throw away any error that happened prior to getting a response.. these will be surfaced to the user
 		// as connection_error status on the response
 		if trace.Response == nil {
 			return call, nil
 		}
+
+		call.ValidJSON = len(trace.ResponseBody) > 0 && json.Valid(trace.ResponseBody)
 
 		return call, err
 	}